Add global config using cli.Config
Just a small package that contains the global config for all commands to use.
The root command will attempt to preload the token from config and set it on the client
automatically, this means that all other commands can just use the client directly
without having to load the token from the config.