cmd/execute: add CLI flags
Changes the execute command to accept CLI flags, the interactive mode
is very nice to have, but it seems like it needs a ton of work & polish
to function well. As an example, we want the interactive mode to immediately
show an error on bad arguments but allow the user to continue from the same
spot.