Add prompt to create missing configs when running `deploy`
This refactors out and re-uses some of `configs set` code to ask the
user to enter in a config value if they try to deploy a task that is
missing configs.

This helps with creating new tasks from e.g. examples (or just when you
want to create a new config). You can run `deploy` and create the
missing configs then.