Refactor execute to accept scripts
Previously execute would accept a yaml definition using the -f flag
since init and deploy removed that flag and just accept a file positional
argument we removed that --file flag.

Essentially changes execute to detect what type of position argument it was
given, it may be a slug, a yaml definition or a script. Once it detects it
it will lookup the task and execute it remotely.

There are still some open questions on how we are going to add local execution
i think it might be a larger effort to add, especially since we rely on a shim etc.

For now, this works and i suggest we merge it and focus on developer experience:

 - Better error messages
 - Nicer CLI UI
 - Edge-cases
 - Detecting root via package.json / .git