Add ap package
The ap package is meant to provide core functionality for the CLI
in the future i would like to move all core CLI types into it and
use other packages for implementations and utilities.

The package has the concept of a Project, unlike airplane.yml or a taskfile (.js|.py)
a project may contain one or more tasks and a framework key that allows us to treat
multiple tasks as one when we deploy or execute.

The package also contains a concept of a Framework, unlike runtime.Interface a framework
may contain one or more tasks, it also knows how to extract commands and maybe run them
using the framework in the future.