Add framework support to init command
Previously the init command accepted a --slug <path>, this adds
the ability to accept --use <framework> which will lookup a framework
and create a "project" file at the root directory "airplane.json".

Unlike `init --slug <path> call`, `init --use <framework>` must be
called from the root directory of the project.

If a file that is required by the framework does not exist an error
is displayed to the user. If the project file already exists a log
is displayed showing the user how to deploy all tasks.