Fix shell to detect Dockerfile.airplane for root
Shell was using Dockerfile.airplane if it happened to be in the root,
but to detect the root itself it was only looking for Dockerfile.
Meaning, unless the root somehow accidentally was right due to another
Dockerfile present or it was the same directory as the script,
Dockerfile.airplane wasn't being used.